split FlowerClient, etc. into fed_kit_train
