+++ android/app/src/main/java/org/eu/fedcampus/android_client/MainActivity.kt
@@ -8,14 +8,13 @@ import androidx.appcompat.app.AppCompatActivity
 import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.launch
 import org.eu.fedcampus.android_client.databinding.ActivityMainBinding
-import org.eu.fedcampus.fed_kit.FlowerClient
 import org.eu.fedcampus.fed_kit.Train
 import org.eu.fedcampus.fed_kit.examples.cifar10.DATA_TYPE
 import org.eu.fedcampus.fed_kit.examples.cifar10.Float3DArray
 import org.eu.fedcampus.fed_kit.examples.cifar10.loadData
 import org.eu.fedcampus.fed_kit.examples.cifar10.sampleSpec
-import org.eu.fedcampus.fed_kit.helpers.deviceId
-import org.eu.fedcampus.fed_kit.helpers.loadMappedFile
+import org.eu.fedcampus.fed_kit_train.FlowerClient
+import org.eu.fedcampus.fed_kit_train.helpers.loadMappedFile
 import java.util.Date
 import java.util.Locale
 
@@ -111,7 +110,7 @@ class MainActivity : AppCompatActivity() {
     suspend fun connectInBackground(participationId: Int, backendUrl: Uri, host: Uri) {
         Log.i(TAG, "Backend URL: $backendUrl")
         train = Train(this, backendUrl.toString(), sampleSpec())
-        train.enableTelemetry(deviceId(this))
+        train.enableTelemetry(org.eu.fedcampus.fed_kit_train.helpers.deviceId(this))
         val modelFile = train.prepareModel(DATA_TYPE)
         val serverData = train.getServerInfo(binding.startFreshCheckBox.isChecked)
         if (serverData.port == null) {
