+++ android/benchmark/build.gradle
@@ -44,7 +44,8 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
-    implementation project(path: ':fed_kit')
 
     // Local dependencies.
+    implementation project(path: ':fed_kit')
+    implementation project(path: ':fed_kit_train')
 }
