+++ android/fed_kit/src/main/java/org/eu/fedcampus/fed_kit/examples/fedmcrnn/fedmcrnn.kt
@@ -1,8 +1,8 @@
 package org.eu.fedcampus.fed_kit.examples.fedmcrnn
 
-import org.eu.fedcampus.fed_kit.SampleSpec
-import org.eu.fedcampus.fed_kit.helpers.maxSquaredErrorLoss
-import org.eu.fedcampus.fed_kit.helpers.placeholderAccuracy
+import org.eu.fedcampus.fed_kit_train.SampleSpec
+import org.eu.fedcampus.fed_kit_train.helpers.maxSquaredErrorLoss
+import org.eu.fedcampus.fed_kit_train.helpers.placeholderAccuracy
 
 fun sampleSpec() = SampleSpec<Float2DArray, FloatArray>(
     { it.toTypedArray() },
