+++ android/fed_kit_train/src/main/java/org/eu/fedcampus/fed_kit_train/db/TFLiteModel.kt
@@ -1,4 +1,4 @@
-package org.eu.fedcampus.fed_kit.db
+package org.eu.fedcampus.fed_kit_train.db
 
 import android.content.Context
 import java.io.File
