+++ fed_kit_client/android/settings.gradle
@@ -12,3 +12,5 @@ apply from: "$flutterSdkPath/packages/flutter_tools/gradle/app_plugin_loader.gra
 
 include ':fed_kit'
 project(':fed_kit').projectDir = new File('../../android/fed_kit/')
+include ':fed_kit_train'
+project(':fed_kit_train').projectDir = new File('../../android/fed_kit_train/')
